@@ -13,10 +13,10 @@
       }
    ],
    "require": {
-      "php": ">=7.1"
+      "php": ">=8.0"
    },
    "require-dev": {
-      "phpunit/phpunit": "7.*"
+      "phpunit/phpunit": "8.*"
    },
    "autoload": {
       "psr-4": {
