@@ -1,10 +1,10 @@
 <?php
 /**
  * @author     Niirrty <niirrty+code@gmail.com>
- * @copyright  © 2017-2020, Niirrty
+ * @copyright  © 2017-2021, Niirrty
  * @package    Niirrty\DynProp
  * @since      2017-10-30
- * @version    0.3.0
+ * @version    0.4.0
  */
 
 
@@ -30,19 +30,19 @@ class ExplicitGetterSetter extends ExplicitGetter
 {
 
 
-    // <editor-fold desc="// – – –   P R O T E C T E D   F I E L D S   – – – – – – – – – – – – – – – – – – – – – –">
+    #region // – – –   P R O T E C T E D   F I E L D S   – – – – – – – – – – – – – – – – – – – – – –
 
     /**
      * The names of all set* methods (without the leading set) that should not be mapped as dynamic properties
      *
      * @type array
      */
-    protected $ignoreSetProperties = [];
+    protected array $ignoreSetProperties = [];
 
-    // </editor-fold>
+    #endregion
 
 
-    // <editor-fold desc="// – – –   P U B L I C   M E T H O D S   – – – – – – – – – – – – – – – – – – – – – – – –">
+    #region // – – –   P U B L I C   M E T H O D S   – – – – – – – – – – – – – – – – – – – – – – – –
 
     /**
      * The magic set method to let you access all setter methods by dynamic properties.
@@ -51,7 +51,7 @@ class ExplicitGetterSetter extends ExplicitGetter
      * @param  mixed  $value   The value to set.
      * @throws \LogicException Is thrown if the property does not exist or if writing is requested but not allowed.
      */
-    public function __set( string $name, $value )
+    public function __set( string $name, mixed $value )
     {
 
         if ( ! $this->hasWritableProperty( $name, $setterName ) )
@@ -84,11 +84,12 @@ class ExplicitGetterSetter extends ExplicitGetter
     /**
      * Returns, if a property with the defined name exists for write access.
      *
-     * @param  string $name       The name of the property.
-     * @param  string $setterName Returns the name of the associated set method, if method returns TRUE.
+     * @param  string     $name       The name of the property.
+     * @param string|null $setterName Returns the name of the associated set method, if method returns TRUE.
+     *
      * @return boolean
      */
-    public function hasWritableProperty( string $name, &$setterName ) : bool
+    public function hasWritableProperty( string $name, ?string &$setterName = null ) : bool
     {
 
         if ( \in_array( $name, $this->ignoreSetProperties ) )
@@ -102,7 +103,7 @@ class ExplicitGetterSetter extends ExplicitGetter
 
     }
 
-    // </editor-fold>
+    #endregion
 
 
 }
